Crate synopsys_usb_otg

USB peripheral driver for Synopsys USB OTG peripherals.

Re-exports

pub use crate::bus::UsbBus;

Modules

bus
USB peripheral driver.

Enums

PhyType
USB PHY type

Traits

UsbPeripheral
A trait for device-specific USB peripherals. Implement this to add support for a new hardware platform. Peripherals that have this trait must have the same register block as STM32 USB OTG peripherals.